Add the generic test suite in build-aux.         This way it's usable easily throughout different projects. 	* tests/test.sh: Move to... 	* build-aux/test.sh: ... here.
git-svn-id: https://svn.gostai.com/svn/common/project-template/trunk@99 f9a80dcf-8710-0410-9714-9a4fb6817a1c

+++ build-aux/test.sh
@@ -80,7 +80,7 @@
 # new_reference_output_file [ret|stdout|stderr] <ref-file>
 new_reference_output_file()
 {
-  test -s $bprog.my_$1 && return 0 # Skip empty files.
+  test -s $bprog.my_$1 || return 0 # Skip empty files.
   # Simply update existing reference output file.
   if test x"$2" != x; then
     cp -f $bprog.my_$1 $2
